test: split report settings bottom margin regression
