+++ tests/CMakeLists.txt
@@ -4645,6 +4645,7 @@ copperfin_add_test(test_visual_asset_editor cf_vfp_assets
     test_visual_asset_editor_method_listing.cpp
     test_visual_asset_editor_property_blank_value.cpp
     test_visual_asset_editor_property_comment_lines.cpp
+    test_visual_asset_editor_report_settings_bottom_margin.cpp
     test_visual_asset_editor_methods.cpp
     test_visual_asset_editor_properties.cpp
     test_visual_asset_editor_report_printer_settings.cpp
